dropout tests in base representation test
trigger ci
